requirements.txt
streamlit>=1.30
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-huggingface>=0.0.1
langchain-pinecone>=0.0.3
langchain-community>=0.0.10
pinecone-client>=3.0
python-dotenv>=1.0
pdfplumber>=0.10
python-docx>=1.1
pydantic>=2.5
spacy>=3.7
scikit-learn>=1.4
plotly>=5.18
sentence-transformers>=2.2.0
